# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "sean",
        "version": "1.0.0",
        "title": "Seans-Python3-Flask-rest-Boilerplate",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [{
        "url": "/"
    }],
...
    "paths": {
...
        "/proprietaire/appart/ask/demarcheur": {
            "post": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Proprietaire request"
                ],
                "summary": "Ask demarcheur",
                "requestBody": {
                    "description": "Ask demarcheur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/demarcheur_ask"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/demarcheur_ask"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/proprietaire/appart/all": {
            "get": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Proprietaire request"
                ],
                "summary": "Get all appart",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/proprietaire/appart/detail": {
            "post": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Proprietaire request"
                ],
                "summary": "Get one appart details",
                "requestBody": {
                    "description": "Get one appart details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/remove_appart"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/remove_appart"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/proprietaire/appart/new": {
            "post": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Proprietaire request"
                ],
                "summary": "Create a new appart",
                "requestBody": {
                    "description": "Create new appart request",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/appart_new"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/appart_new"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/proprietaire/appart/update": {
            "post": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Proprietaire request"
                ],
                "summary": "Update appart detail",
                "requestBody": {
                    "description": "Update appart detail",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/appart_update"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/appart_update"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/proprietaire/appart/remove": {
            "delete": {
                "parameters": [{
                    "name": "appart_id",
                    "in": "query",
                    "required": false,
                    "description": "l'id de l'appartement",
                    "type": "string"
                }],
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Proprietaire request"
                ],
                "summary": "remove appart detail",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/remove_appart"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "remove_appart": {
                "type": "object",
                "properties": {
                    "appart_id": {
                        "type": "string",
                        "format": "string"
                    }
                }
            },
            "appart_new": {
                "type": "object",
                "properties": {
                    "certified": {
                        "type": "bool",
                        "format": "string",
                        "example": false
                    },
                    "demarcheur_id": {
                        "type": "string",
                        "format": "string"
                    },
                    "caption": {
                        "type": "string",
                        "format": "string"
                    },
                    "country": {
                        "type": "string",
                        "format": "string"
                    },
                    "town": {
                        "type": "string",
                        "format": "string"
                    },
                    "district": {
                        "type": "string",
                        "format": "string"
                    },
                    "description": {
                        "type": "string",
                        "format": "string"
                    },
                    "specifications": {
                        "type": "object",
                        "format": "object",
                        "example": {
                            "room": 1,
                            "shower": 1,
                            "toilet": 1,
                            "living room": 1,
                            "kitchen": 1,
                            "veranda": 1,
                            "backyard": 1,
                            "garage": 1
                        }
                    },
                    "price": {
                        "type": "string",
                        "format": "string",
                        "example": "25.000 fcfa"
                    },
                    "is_occupied": {
                        "type": "bool",
                        "format": "string",
                        "example": false
                    },
                    "rent": {
                        "type": "bool",
                        "format": "string",
                        "example": true
                    },
                    "co-rent": {
                        "type": "bool",
                        "format": "string",
                        "example": true
                    },
                    "interested_in_co_renting": {
                        "type": "array",
                        "format": "array",
                        "example": ["id_user"]
                    },
                    "filenames": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        }
                    }
                }
            },
            "appart_update": {
                "type": "object",
                "properties": {
                    "appart_id": {
                        "type": "string",
                        "format": "string"
                    },
                    "demarcheur_id": {
                        "type": "string",
                        "format": "string"
                    },
                    "caption": {
                        "type": "string",
                        "format": "string"
                    },
                    "country": {
                        "type": "string",
                        "format": "string"
                    },
                    "town": {
                        "type": "string",
                        "format": "string"
                    },
                    "district": {
                        "type": "string",
                        "format": "string"
                    },
                    "description": {
                        "type": "string",
                        "format": "string"
                    },
                    "specifications": {
                        "type": "object",
                        "format": "object",
                        "example": {
                            "room": 1,
                            "shower": 1,
                            "toilet": 1,
                            "living room": 1,
                            "kitchen": 1,
                            "veranda": 1,
                            "backyard": 1,
                            "garage": 1
                        }
                    },
                    "price": {
                        "type": "string",
                        "format": "string",
                        "example": "25.000 fcfa"
                    },
                    "is_occupied": {
                        "type": "bool",
                        "format": "string",
                        "example": false
                    },
                    "rent": {
                        "type": "bool",
                        "format": "string",
                        "example": true
                    },
                    "co-rent": {
                        "type": "bool",
                        "format": "string",
                        "example": true
                    },
                    "interested_in_co_renting": {
                        "type": "array",
                        "format": "array",
                        "example": ["id_user"]
                    }
                }
            },
            "demarcheur_ask": {
                "type": "object",
                "properties": {
                    "demarcheur_id": {
                        "type": "string",
                        "format": "string"
                    },
                    "appart_id": {
                        "type": "string",
                        "format": "string"
                    }
                }
            },
...
        }
    }
}